pi_instagram_stack/requirements.txt
# =============================================
# pi_instagram_stack — Python Dependencies
# =============================================
# Raspberry Pi 5 / ARM64 compatible

# --- Google Gemini AI ---
google-generativeai>=0.8.0

# --- Database ---
psycopg2-binary>=2.9.9

# --- Data Models ---
pydantic>=2.5.0

# --- Environment ---
python-dotenv>=1.0.0

# --- HTTP ---
requests>=2.31.0

# --- News Scraping ---
feedparser>=6.0.0

# --- Video Download ---
yt-dlp>=2024.1.0

# --- Video Assembly ---
moviepy>=1.0.3
Pillow>=10.0.0

# --- Logging ---
colorlog>=6.7.0
